Parse client version out of client.cfg
Add the client version to client.cfg and load
it from that instead of from client.cuo. This is one step toward
making orion.dll and client.cuo optional.

The user can simply provide the string version number and the
client will convert that to an integer that can be compared
to the enums of important client version cutoffs. This is a little
tricky because UO changed versioning schemes half way through its
life. Versions come in two forms:

1.2.3a
1.2.3.4

To represent these both as an integer, treat the appearance of the
non digit as the fourth element in the version, where a is 0, be
is 1, etc.