client.cfg now understands ClientPath
Add an entry to client.cfg, ClientPath. ClientPath can point
at a regular UO install. This allows the OrionUO client to load
media files from an existing UO install from a different
directory without modifying it.

This is a big step toward making it easier to install and use
OrionUO.