Add cotire for speeding up compilation
Add two project modes, with libraries or without. This option is
configured via CMake ORION_BUILD_MODULES, currently set to off because
it does not work well with cotire unity builds.

Giving preference to unity builds as it is fast, currently building in
about 12s. Can be further optimized if fixing all issues with modules
and stdafx.