+ Bugfix: Character selection screen with empty slots
ReadString* fix to trim strings on null char, this was causing the empty
slots and many things elsewhere to not work correctly as the internal
storage size was different of the internal C-string representation size.

Fix ApplyInstall to write a sized buffer of bytes for now until we can
rewrite ReadString/WriteString in something safer to misuse.

Little cleanup on WispDataStream.